This function downloads a directory of model runs to a local directory from the Open Science Grid (OSG).

Options exist for cleaning the local and remote directory after download

Usage

osg_download_ss_dir(
  session = NULL,
  unix_name,
  login_node,
  rsa_keyfile = NULL,
  rsa_passphrase = NULL,
  remote_dir_stem,
  remote_dirs,
  download_dir_stem,
  files_to_download = c("End.tar.gz"),
  untar_local = TRUE,
  clean_remote = TRUE,
  delete_remote = FALSE,
  verbose = TRUE
)

Arguments

session

ssh connection created by osg_connect.

unix_name

Character string giving OSG unix login name.

login_node

Character string giving OSG login node (e.g., login05.osgconnect.net).

rsa_keyfile

Path to private key file. Must be in OpenSSH format (see details). Default is NULL. See ssh_connect for more details.

rsa_passphrase

Either a string or a callback function for password prompt. Default is NULL. See ssh_connect for more details.

remote_dir_stem

Path on OSG login node to directory to download.

remote_dirs

Character vector of sub-directories to download from remote_dir_stem.

download_dir_stem

Path to directory on local machin to download into.

files_to_download

Files to download from each directory given by remote_dirs.

untar_local

Boolean denoting whether or not to untar all downloaded *tar.gz files.

clean_remote

Boolean denoting whether or not to clean remote_dirs using osg_clean.

delete_remote

Boolean denoting whether or not to delete remote_dirs.

verbose

Boolean denoting if function details should be printed.

Value

Returns 0 on exit.