This function writes an R script which can be executed within the wrapper.sh to set-up and run diagnostics using existing r4ss code.

This function saves a copy locally, uploads a copy to Open Science Grid (OSG), and modifies permissions so that it can be executed remotely

Usage

osg_r_create(
  session = NULL,
  unix_name,
  login_node,
  rsa_keyfile = NULL,
  rsa_passphrase = NULL,
  local_shell_path = NULL,
  remote_shell_path = "scripts/r/",
  diagnostic_type = NULL,
  overwrite = TRUE,
  verbose = TRUE,
  retro_n_years = NULL,
  r0_maxdiff = NULL,
  r0_step = NULL
)

Arguments

session

ssh connection created by osg_connect.

unix_name

Character string giving OSG unix login name.

login_node

Character string giving OSG login node (e.g., login05.osgconnect.net).

rsa_keyfile

Path to private key file. Must be in OpenSSH format (see details). Default is NULL. See ssh_connect for more details.

rsa_passphrase

Either a string or a callback function for password prompt. Default is NULL. See ssh_connect for more details.

local_shell_path

Path to directory where condor_submit script is written. Defaults to tempdir.

remote_shell_path

Path to directory on OSG login node where condor_submit script is written.

diagnostic_type

A character string specifying which diagnostic to run.

"01_run_retro"

Conduct a retrospective analysis of length retro_n_years.

"02_run_R0profile"

Conduct a likelihood profile on R0 where the profile extends +/- r0_maxdiff from the log(R0) maximum likelihood estimate at an interval specified by r0_step.

"03_run_aspm"

Conduct age structured production (ASPM) and deterministic recruitment runs of the model. Warning: for the ASPM there are some hard coded settings that may not be appropriate for your model.

overwrite

If the file given by file_name exists in local_shell_path or remote_shell_path overwrite if TRUE.

verbose

Boolean denoting if function details should be printed.

retro_n_years

The number of years for the retrospective analysis.

r0_maxdiff

The maximum distance from the log(R0) maximum likelihood estimate that the profile will extend.

r0_step

The step size used in profiling (on a log scale).

Value

Returns 0 on exit.