This function uploads a directory structure to the Open Science Grid (OSG).

This function uploads a directory structure to the Open Science Grid (OSG).

osg_upload_ss_dir(
  session = NULL,
  unix_name,
  login_node,
  rsa_keyfile = NULL,
  rsa_passphrase = NULL,
  local_dir_path,
  local_dir_names = NULL,
  remote_dir_path,
  files_to_upload,
  target_dir_path,
  target_dir_txt_name,
  verbose = TRUE
)

Arguments

session

ssh connection created by osg_connect.

unix_name

Character string giving OSG unix login name.

login_node

Character string giving OSG login node (e.g., login05.osgconnect.net).

rsa_keyfile

Path to private key file. Must be in OpenSSH format (see details). Default is NULL. See ssh_connect for more details.

rsa_passphrase

Either a string or a callback function for password prompt. Default is NULL. See ssh_connect for more details.

local_dir_path

Path to directory on local machine to upload to OSG.

local_dir_names

Character vector of sub-directories within local_dir_path to upload to OSG.

remote_dir_path

Path on OSG login node to upload local_dir_names into.

files_to_upload

Character vector giving file names to upload to OSG.

target_dir_path

Path on OSG login node with text file giving each directory to execute condor jobs in.

target_dir_txt_name

File name given to text file containing path of each directory to execute condor jobs in.

verbose

Boolean denoting if function details should be printed.

Value

Returns 0 on exit.